pyproject.toml
[project]
name = "buscamidepto-scrapping"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles==24.1.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.18",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anthropic==0.51.0",
    "anyio==4.9.0",
    "appnope==0.1.4 ; sys_platform == 'darwin'",
    "asttokens==3.0.0",
    "attrs==25.3.0",
    "backoff==2.2.1",
    "beautifulsoup4==4.13.4",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    "chardet==5.2.0",
    "charset-normalizer==3.4.2",
    "click==8.2.0",
    "colorama==0.4.6 ; sys_platform == 'win32'",
    "comm==0.2.2",
    "cryptography==44.0.3",
    "dataclasses-json==0.6.7",
    "debugpy==1.8.14",
    "decorator==5.2.1",
    "distro==1.9.0",
    "emoji==2.14.1",
    "executing==2.2.0",
    "fastapi==0.115.12",
    "filetype==1.2.0",
    "frozenlist==1.6.0",
    "greenlet==3.2.2 ; (python_full_version < '3.14' and platform_machine == 'AMD64') or (python_full_version < '3.14' and platform_machine == 'WIN32') or (python_full_version < '3.14' and platform_machine == 'aarch64') or (python_full_version < '3.14' and platform_machine == 'amd64') or (python_full_version < '3.14' and platform_machine == 'ppc64le') or (python_full_version < '3.14' and platform_machine == 'win32') or (python_full_version < '3.14' and platform_machine == 'x86_64')",
    "h11==0.16.0",
    "html5lib==1.1",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "idna==3.10",
    "iniconfig==2.1.0",
    "ipykernel==6.29.5",
    "ipython==9.2.0",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "jiter==0.9.0",
    "joblib==1.5.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jupyter-client==8.6.3",
    "jupyter-core==5.7.2",
    "langchain==0.3.25",
    "langchain-anthropic==0.3.13",
    "langchain-community==0.3.24",
    "langchain-core==0.3.60",
    "langchain-ollama==0.3.3",
    "langchain-openai==0.3.17",
    "langchain-text-splitters==0.3.8",
    "langdetect==1.0.9",
    "langgraph==0.4.5",
    "langgraph-checkpoint==2.0.26",
    "langgraph-prebuilt==0.1.8 ; python_full_version < '4.0'",
    "langgraph-sdk==0.1.69 ; python_full_version < '4.0'",
    "langsmith==0.3.42",
    "lxml==5.4.0",
    "mangum==0.19.0",
    "marshmallow==3.26.1",
    "matplotlib-inline==0.1.7",
    "multidict==6.4.3",
    "mypy-extensions==1.1.0",
    "nest-asyncio==1.6.0",
    "nltk==3.9.1",
    "numpy==2.2.5",
    "olefile==0.47",
    "ollama==0.4.8",
    "openai==1.78.1",
    "orjson==3.10.18 ; python_full_version < '4.0' or platform_python_implementation != 'PyPy'",
    "ormsgpack==1.9.1",
    "outcome==1.3.0.post0",
    "packaging==24.0",
    "parso==0.8.4",
    "pexpect==4.9.0 ; sys_platform != 'emscripten' and sys_platform != 'win32'",
    "platformdirs==4.3.8",
    "pluggy==1.6.0",
    "prompt-toolkit==3.0.51",
    "propcache==0.3.1",
    "psutil==7.0.0",
    "ptyprocess==0.7.0 ; sys_platform != 'emscripten' and sys_platform != 'win32'",
    "pure-eval==0.2.3",
    "pycparser==2.22",
    "pydantic==2.11.4",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.9.1",
    "pygments==2.19.1",
    "pypdf==5.5.0",
    "pysocks==1.7.1",
    "pytest==8.3.5",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python-iso639==2025.2.18",
    "python-magic==0.4.27",
    "python-oxmsg==0.0.2",
    "pywin32==310 ; platform_python_implementation != 'PyPy' and sys_platform == 'win32'",
    "pyyaml==6.0.2",
    "pyzmq==26.4.0",
    "rapidfuzz==3.13.0",
    "regex==2024.11.6",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "selenium==4.32.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "soupsieve==2.7",
    "sqlalchemy==2.0.41",
    "stack-data==0.6.3",
    "starlette==0.46.2",
    "tavily-python==0.7.2",
    "tenacity==9.1.2",
    "tiktoken==0.9.0",
    "tornado==6.5",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "trio==0.30.0",
    "trio-websocket==0.12.2",
    "typing-extensions==4.13.2",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.0",
    "unstructured==0.17.2",
    "unstructured-client==0.35.0",
    "urllib3==2.4.0",
    "uvicorn==0.34.2",
    "wcwidth==0.2.13",
    "webdriver-manager==4.0.2",
    "webencodings==0.5.1",
    "websocket-client==1.8.0",
    "wrapt==1.17.2",
    "wsproto==1.2.0",
    "xxhash==3.5.0",
    "yarl==1.20.0",
    "zstandard==0.23.0",
]

[dependency-groups]
dev = [
    # In development mode, include the FastAPI development server.
    "fastapi[standard]>=0.115",
]
